Fix: program no longer crashes

+++ fileprocessing.h
@@ -9,8 +9,9 @@
 #include "imagegeneration.h"
 
 
-void  processFilterFile(const char* filename, short int* parameter, short int* parameters, short int* flag);
+void processFilterFile(const char* filename, short int* parameter, short int* parameters, int* flag);
 void  writeFile(FILE* archivo, t_pixel** mat, size_t filas, size_t columnas);
 char* generateImageName(const char* filter,  char* fileName);
 int   readHeader(FILE* file, t_header *header);
+
 #endif // FILE_PROCESSING_INCLUDED
